* Flipped TypeScript Int back to Float to avoid typescript not allowing larger Int types * Refactored create_options to be its own function with tests and reduced types around it

+++ x-pack/plugins/secops/common/graphql/shared/schema.gql.ts
@@ -23,7 +23,7 @@ export const sharedSchema = gql`
 
   input PaginationInput {
     "The limit parameter allows you to configure the maximum amount of items to be returned"
-    limit: Int!
+    limit: Float!
     "The cursor parameter defines the next result you want to fetch"
     cursor: String
     "The tiebreaker parameter allow to be more precise to fetch the next item"
